refactor(sentry): move into module
