--- src/server/controllers/interfaces/SentryControllerInterface.ts
@@ -1,8 +1,0 @@
-import Express from 'express'
-
-export interface SentryControllerInterface {
-  /**
-   * Requests for the Sentry DNS.
-   */
-  getSentryDns(req: Express.Request, res: Express.Response): void
-}
