+++ src/server/inversify.config.ts
@@ -37,7 +37,7 @@ import { StatisticsRepository } from './repositories/StatisticsRepository'
 import { StatisticsService } from './services/StatisticsService'
 import { StatisticsController } from './controllers/StatisticsController'
 import { RotatingLinksController } from './controllers/RotatingLinksController'
-import { SentryController } from './controllers/SentryController'
+import { SentryController } from './modules/sentry/SentryController'
 import { LoginController } from './controllers/LoginController'
 import { AuthService } from './services/AuthService'
 import { LogoutController } from './controllers/LogoutController'
