+++ src/server/modules/sentry/SentryController.ts
@@ -1,11 +1,10 @@
 /* eslint-disable class-methods-use-this */
 import Express from 'express'
 import { injectable } from 'inversify'
-import { sentryDns } from '../config'
-import { SentryControllerInterface } from './interfaces/SentryControllerInterface'
+import { sentryDns } from '../../config'
 
 @injectable()
-export class SentryController implements SentryControllerInterface {
+export class SentryController {
   getSentryDns(_: Express.Request, res: Express.Response) {
     res.send(sentryDns)
     return
