+++ src/server/modules/sentry/index.ts
@@ -1,0 +1,5 @@
+import { SentryController } from './SentryController'
+
+export { SentryController } from './SentryController'
+
+export default SentryController
